# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FERNS — Federated Ecological Resource Network System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lady-bird-johnson:
    get:
      operationId: getLadyBirdJohnson
      tags: [lady-bird-johnson]
      summary: Verify a Lady Bird Johnson Wildflower Center species profile URL from a USDA Plants symbol
      description: >
        Accepts a USDA Plants symbol (e.g. TRGI for Trillium grandiflorum) and returns a verified
        direct species profile URL at https://www.wildflower.org/plants/result.php?id_plant={SYMBOL}.
        Verification is performed via HTTP GET with redirect:manual and a browser-like User-Agent
        (required — the site returns 403 to generic agents). HTTP 200 = found; 3xx redirect = not_found;
        5xx/network = unverified (not cached). Results are cached in lbj_url_cache (90-day TTL when
        found, 30-day when not_found). Obtain USDA Plants symbols via /usda-plants.
      parameters:
        - name: usda_symbol
          in: query
          required: true
          description: USDA Plants symbol (e.g. TRGI for Trillium grandiflorum). Obtain via /usda-plants.
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Lady Bird Johnson profile URL verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LbjUrlCheckResponse"
        "400":
          description: Missing or invalid usda_symbol parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lady-bird-johnson/species-text:
    get:
      operationId: getLadyBirdJohnsonSpeciesText
      tags: [lady-bird-johnson]
      summary: Scrape and return species page text from Lady Bird Johnson Wildflower Center
      description: >
        Fetches and caches botanical prose text from the Lady Bird Johnson Wildflower Center
        (wildflower.org/plants) species profile page identified by a USDA Plants symbol.
        Verification and scraping use a browser-like User-Agent (required — site returns 403
        to generic agents) with redirect:manual to detect not-found responses (3xx redirect).
        Sections extracted include all h3-delimited prose; "Find Seeds or Plants" and
        "Mr. Smarty Plants says" sections are excluded. First call fetches live; subsequent
        calls return the cached text. Use ?refresh=true to force a re-scrape. Returns
        found=false when the symbol is not found on the Wildflower Center site.
        Obtain USDA Plants symbols via /usda-plants.
      parameters:
        - name: usda_symbol
          in: query
          required: true
          description: USDA Plants symbol (e.g. TRGI for Trillium grandiflorum). Obtain via /usda-plants.
          schema:
            type: string
            minLength: 1
        - name: refresh
          in: query
          required: false
          description: If "true", bypass cache and re-scrape the live page
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: Species page text result (found or not-found — check the found field)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LbjSpeciesTextResponse"
        "400":
          description: Missing symbol parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
    FernsProvenance:
      type: object
      description: >
        Provenance block present on every FERNS API response.
        Identity fields (source_id, fetched_at, method, upstream_url) are always present.
        Text fields (general_summary, technical_details) are conditionally present
        based on the provenance_verbosity query parameter (full|summary|none).
      required:
        - source_id
        - fetched_at
        - method
        - upstream_url
      properties:
        source_id:
          type: string
          description: Stable identifier for this data source (e.g. bonap-napa)
        fetched_at:
          type: string
          format: date-time
          description: When this record was obtained from the source
        method:
          type: string
          description: "How the data was obtained: api_fetch | blob_import | llm_synthesis"
        upstream_url:
          type: string
          description: Where this data came from (API endpoint, file path, or registry entry)
        general_summary:
          type: string
          description: Plain language description readable by a homeowner or community member
        technical_details:
          type: string
          description: >
            Research-grade description: methods, measurement protocols, algorithms,
            citations, and transformations — sufficient for a scientist to evaluate and reproduce
        matched_input:
          type: string
          description: >
            The normalized input that was actually used for this lookup (e.g., the name
            as queried). Present on endpoints that accept a name parameter.
...
    LbjUrlCheckResponse:
      type: object
      description: >
        Response from /lady-bird-johnson — profile URL verification result keyed by USDA Plants symbol.
      required:
        - found
        - queried_at
      properties:
        found:
          type: boolean
          description: True when HTTP 200 returned for the profile URL (status="found").
        queried_at:
          type: string
          format: date-time
        source_url:
          type: string
        provenance:
          $ref: "#/components/schemas/FernsProvenance"
        data:
          type: object
          properties:
            usda_symbol:
              type: string
              description: USDA Plants symbol (uppercased).
            profile_url:
              type: string
              nullable: true
              description: Direct profile URL when found; null when not_found or unverified.
            status:
              type: string
              enum: [found, not_found, unverified]
              description: >
                "found" — HTTP 200 returned. "not_found" — 3xx redirect or 4xx returned.
                "unverified" — 5xx or network error; result not cached.
            found:
              type: boolean
              description: True when status is "found".
            http_status:
              type: integer
              nullable: true
              description: HTTP status code returned by the verification request.
            validation_method:
              type: string
              description: Always "http_get_manual_redirect" for this endpoint.
            verified_at:
              type: string
              format: date-time
              nullable: true
              description: >
                Timestamp when the verification HTTP request was made.
                Null when status is "unverified" (network/5xx error prevented caching).
            cache_hit:
              type: boolean
              description: True if the result was served from lbj_url_cache.
    LbjSpeciesTextResponse:
      type: object
      description: >
        Response from /lady-bird-johnson/species-text — scraped species page text keyed by USDA Plants symbol.
      required:
        - found
        - queried_at
        - cache_status
      properties:
        found:
          type: boolean
          description: True if the species page was found and text was extracted.
        queried_at:
          type: string
          format: date-time
        source_url:
          type: string
        cache_status:
          type: string
          enum: [hit, miss]
          description: >
            "hit" — returned from species_page_text_cache. "miss" — live fetch performed.
        fetch_error:
          type: string
          description: >
            Present only when a transient upstream error prevented the scrape. Not cached.
        scraped_at:
          type: string
          format: date-time
          description: Timestamp of when the text was originally scraped.
        provenance:
          $ref: "#/components/schemas/FernsProvenance"
        data:
          nullable: true
          type: object
          description: Present when found=true. Null when not found.
          properties:
            usda_symbol:
              type: string
              description: USDA Plants symbol (uppercased).
            url:
              type: string
              description: The profile URL that was scraped.
            sections:
              type: object
              nullable: true
              additionalProperties:
                type: string
              description: >
                Labeled prose sections extracted from the page (h3-delimited).
                "Find Seeds or Plants" and "Mr. Smarty Plants says" sections are excluded.
            full_text:
              type: string
              nullable: true
              description: All sections concatenated as Label-colon-text blocks, separated by double newlines.
...